requirements.txt
fastapi
uvicorn
python-multipart
transformers
torchaudio
whisper


# Run this below to start the server
# uvicorn main:app --reload